+++ cmd/fort/main.go
@@ -14,7 +14,7 @@ import (
 	"github.com/djadmin/fort/internal/checks"
 )
 
-var version = "0.2.0"
+var version = "0.3.0"
 
 func main() {
 	var (